Add Kafka-to-HTTP bridge.
Our Nitro enclave is unable to talk to our Kafka cluster directly
because proxied communication between the enclave and the rest of the
world is limited to HTTP -- we rely on Go's HTTP(S)_PROXY environment
variable.

The purpose of this tool is to bridge HTTP requests to Kafka messages.